improve performance of the generator
Instead of using bruteforce to generate the offsets for each timezone,
query NodaTime for the end of the timezone-interval.

This actually changes one single entry in the timezone_database.h, for Africa/Monrovia,
but this is because the previous entry was incorrect since it isn't at a clean minute
border and thus the correct timestamp for the interval was skipped.

This actually raises another issue, the offset for Africa/Monrovia before 1972 12:44:30 AM UTC is
actually incorrect right now because it's stored as a 44 minutes, but correct would be 44:30 minutes.